login page updated moved to react
