+++ final-proj/package.json
@@ -18,6 +18,7 @@
     "passport-github2": "^0.1.12",
     "react": "^19.1.1",
     "react-dom": "^19.1.1",
+    "react-router-dom": "^7.9.4",
     "vite-express": "^0.21.1",
     "ws": "^8.18.3"
   },
