--- final-proj/public/login.html
@@ -1,47 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <title>Sign in</title>
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet">
-    <script src="/main.js" defer></script>
-  </head>
-  <body class="bg-light">
-    <div class="container min-vh-100 d-flex align-items-center">
-      <div class="row w-100 justify-content-center">
-        <div class="col-12 col-sm-10 col-md-8 col-lg-5">
-          <div class="card shadow-sm">
-            <div class="card-body p-4 p-sm-5">
-              <h1 class="h3 mb-3 text-center">Sign in</h1>
-              <p class="text-secondary text-center mb-4">Use email and password to login/create and account or continue with GitHub.</p>
-              <div id="err" class="alert alert-danger d-none" role="alert"></div>
-              <div id="ok" class="alert alert-success d-none" role="alert"></div>
-              <form id="loginForm" novalidate class="needs-validation">
-                <div class="mb-3">
-                  <label for="email" class="form-label">Email</label>
-                  <input id="email" name="email" type="email" class="form-control" autocomplete="email" required>
-                </div>
-                <div class="mb-3">
-                  <label for="password" class="form-label">Password</label>
-                  <div class="input-group">
-                    <input id="password" name="password" type="password" class="form-control" autocomplete="current-password" required>
-                  </div>
-                </div>
-                <button type="submit" class="btn btn-primary w-100">Sign in</button>
-              </form>
-              <div class="d-flex align-items-center my-4">
-                <hr class="flex-grow-1">
-                <span class="px-2 text-secondary">or</span>
-                <hr class="flex-grow-1">
-              </div>
-              <a class="btn btn-dark w-100" href="/auth/github" role="button">
-                Continue with GitHub
-              </a>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  </body>
-</html>
